Add Build Mod.ps1
This commit adds the Build Mod.ps1 PowerShell script to the repository. This script builds and compresses the mod, which means cloning the repository to a temporary folder, setting up the mod together with its optional features, and finally compressing the build to a 7z file (optionally). Usually, you would only need to have git installed in your system/user path and set the global paths correctly to make the script build the mod.